[Metric] Model Exporter Done.
`rpc.py` is modified so that the metric monitor is put in. Metric
exporter now run along side with the model closure and communicate
via a pipe. The prometheus client is updated whenever a new replica
is added.

`docker_metric_utils.py` is the file for all metric related code that
is put in `docker_container_manager.py` to reduce clutter. Prometheus
listen to port 1390, a port specified in `container_manager.py` as a
constant.

The requirement for `prometheus_client` package is added to three
places: `requirements.txt`, `setup.py`, `RPCDockerfile`

I still need to work on integration test though.

<Simon Mo>