Metrics Monitoring: A minimal viable product (#348)
* [Metrics] A working node exporter done.

User can now just run clipper like regular run and spin up a
metric monitor under the condition:
1. There is a clipper exporter docker image.

<Simon Mo>

* [Metric] Model Exporter Done.

`rpc.py` is modified so that the metric monitor is put in. Metric
exporter now run along side with the model closure and communicate
via a pipe. The prometheus client is updated whenever a new replica
is added.

`docker_metric_utils.py` is the file for all metric related code that
is put in `docker_container_manager.py` to reduce clutter. Prometheus
listen to port 1390, a port specified in `container_manager.py` as a
constant.

The requirement for `prometheus_client` package is added to three
places: `requirements.txt`, `setup.py`, `RPCDockerfile`

I still need to work on integration test though.

<Simon Mo>

* [Metric] Add FrontendExporter Docker image

Changed the corresponding files for the frontend exporter to be built.
- `build_docker_image.sh` and `docker_metric_utils.py` agree on the name
'frontend-exporter'
- python file and Dockerfile are in place

<Simon Mo>

* [Metric] Add docstrings

* [Metric] Add integration test for clipper metric

This is just a simple modification of basic_query example.
It tests for two replicas to see if the prometheus client
is monitoring all three node exporter.

* [Metric] Format Code

* [Metrics] Bug Fix, update the name accoridingly

* [Metric] Small Fixes

Update the code according to Dan's comments. Moved the front-end
container to a separate directory.

* [Metric] Format Code

* [Metric] Skipped Metrics to Let Jenkins Build Images

* [Metric] Revert the files; docker imgs built before test

* Move the comments; restart the tests

* [Metric]  Add version tag to the frontend-exporter

* [Metric] Add clipper_metric to run_unittests.sh

* [Metric] Trigger another CI check

Pretty sure it isn't the issue with my code. But it now appears
common enough.

* [Metrics] A working node exporter done.

User can now just run clipper like regular run and spin up a
metric monitor under the condition:
1. There is a clipper exporter docker image.

<Simon Mo>

* [Metric] Model Exporter Done.

`rpc.py` is modified so that the metric monitor is put in. Metric
exporter now run along side with the model closure and communicate
via a pipe. The prometheus client is updated whenever a new replica
is added.

`docker_metric_utils.py` is the file for all metric related code that
is put in `docker_container_manager.py` to reduce clutter. Prometheus
listen to port 1390, a port specified in `container_manager.py` as a
constant.

The requirement for `prometheus_client` package is added to three
places: `requirements.txt`, `setup.py`, `RPCDockerfile`

I still need to work on integration test though.

<Simon Mo>

* [Metric] Add FrontendExporter Docker image

Changed the corresponding files for the frontend exporter to be built.
- `build_docker_image.sh` and `docker_metric_utils.py` agree on the name
'frontend-exporter'
- python file and Dockerfile are in place

<Simon Mo>

* [Metric] Add docstrings

* [Metric] Add integration test for clipper metric

This is just a simple modification of basic_query example.
It tests for two replicas to see if the prometheus client
is monitoring all three node exporter.

* [Metric] Format Code

* [Metrics] Bug Fix, update the name accoridingly

* [Metric] Small Fixes

Update the code according to Dan's comments. Moved the front-end
container to a separate directory.

* [Metric] Format Code

* [Metric] Skipped Metrics to Let Jenkins Build Images

* [Metric] Revert the files; docker imgs built before test

* Move the comments; restart the tests

* [Metric]  Add version tag to the frontend-exporter

* [Metric] Add clipper_metric to run_unittests.sh

* [Metric] Trigger another CI check

Pretty sure it isn't the issue with my code. But it now appears
common enough.

* Address comments, fix typo

* [Metric] change the redis port for integration-test

So that it won't clash with other testing progress