Fix a bug that a new model replica couldn't be registerd with the Clipper's RPCService
* problem in detail
I suspect that in rare case, the 'unique' connection id generated automatically by new ZMQ connection between model replica and query_frontend replica, was the same with previous ids, so query_frontend's RPCService couldn't register the model container with it, and then this error was occurred finally.

* Error logs
```
[00:36:51.032][info] [RPC] Found message to receive
[00:36:51.034][info] [RPC] Found message to receive
[00:36:51.036][info] [RPC] Found message to receive
terminate called after throwing an instance of 'std::runtime_error'
 what(): Documenting receive time for an unregistered container
```

* how to resolve
We have to remove the connection related with an inactive container from the 'connections' bimap.